refactor(schema): relocate SchemaIdentOutput to processor-schema; strip streamlib-moq facade dep (#1263)
Precursor library-side change that unblocks the engine-free flip of the moq
package. `SchemaIdentOutput` (+ its embedded `SemanticVersionOutput`) moves out
of the engine (`core/json_schema.rs`) into the engine-free
`streamlib-processor-schema` crate, which already owns `SchemaIdent` / `SemVer`.
The engine re-exports both so `core::json_schema::*` and the
`streamlib::sdk::json_schema` facade the api-server consumes resolve unchanged.

The DTO stays serde + schemars in the plugin authoring chain; a `utoipa::ToSchema`
derive is added behind an off-by-default `openapi` feature that the engine
enables for its API layer. A plugin `.slpkg` built engine-free reaches the DTO
through processor-schema without compiling the host-only OpenAPI machinery.

`runtime/streamlib-moq` drops its `streamlib` facade dep and retargets to the
granular engine-free crates (`streamlib-error`, `streamlib-processor-schema`).
A package cdylib links this transport lib, so the facade here would statically
bundle a second copy of streamlib-engine into the host — a plugin-safety
correctness requirement the direct-dep grep in check-boundaries does not catch.
